build: buildSrc -> build-logic 包含构建 (阶段 5)
included build 不像 buildSrc 那样把自己的 classpath 泄漏给主构建的 script classpath,
因此这一步的绝大部分工作是把原本靠"泄漏"才能解析的插件请求补成显式声明。

结构:
- git mv buildSrc build-logic, 新增 build-logic/settings.gradle.kts
  (included build 不共享主构建的 repositories / 版本目录, 都要自己声明一遍)
- settings.gradle.kts: pluginManagement { includeBuild("build-logic") }
- 根脚本删掉整个 `plugins { alias(...) apply false }` 与 buildscript {}
  —— 那串 apply false 的唯一作用就是把插件塞进根 buildscript classpath 供子项目继承,
  是 buildSrc 时代的补丁
- build-logic/build.gradle.kts 显式读仓库根的 gradle.properties:
  included build 不继承主构建的 gradle.properties, 否则 jvm.toolchain.* 会悄悄回落到默认 JDK

补齐的插件版本 (gradle/libs.versions.toml):
- kotlin-multiplatform / kotlin-jvm / android-* / jetbrains-compose /
  kotlin-plugin-compose / kotlin-native-cocoapods / kotlinx-atomicfu 全部补上 version.ref
- kotlin-parcelize 是唯一例外, 用【完整 id + 不带版本】:
  短 id `kotlin-parcelize` 在 classpath 上没有注册 (报 not found),
  带版本又会撞上 "already on the classpath with an unknown version"

新增约定插件 (让插件解析不再依赖 classpath 泄漏):
- ani.jvm-library      (7 个模块): 由它应用 kotlin.jvm。模块自己 alias 会在
                       serialization 等 KGP 子插件先解析后撞 "already on the classpath"
- ani.android-library  (2 个) / ani.android-application (1 个): 应用 AGP + parcelize。
                       不应用 kotlin.android —— AGP 9 起 Kotlin 支持内建, 显式应用会被 KGP 拒绝
- flatten-source-sets 改名 ani.flatten-source-sets, 调用方从 `反引号` 改成 id("..."),
  included build 的插件没有反引号访问器

顺带修复 :app:shared:app-data 的 Room 配置:
它同时用 room { schemaDirectory(...) } 和 ksp { arg("room.schemaLocation", ...) },
而 Room 插件明确禁止两者并存。以前不报错是因为 androidx.room 由【根项目的 buildscript
classloader】加载 (子项目 classpath 里根本看不到它, 已用 buildEnvironment 对比确认),
冲突检查没生效; build-logic 化后每个模块自己解析插件, 检查才真正跑起来。
已删除重复的 ksp arg, schema 输出目录不变。

验证 (基线取自 26c2b72d9 的 pristine worktree):
- 项目拓扑不变
- 5 个模块的 (configuration, 解析后 module) 集合与基线逐条相同, 合计 28729 条
- compileKotlinDesktop + 三个 metadata 任务: BUILD SUCCESSFUL (631 tasks)
- desktopTest: BUILD SUCCESSFUL
- compileDefaultDebugKotlin + compileDefaultReleaseKotlin: BUILD SUCCESSFUL
- 配置缓存 store + reuse 正常
- F4 收益实测: 改动 build-logic 源码后, 需要重编的主构建 build script 数量为 0
  (buildSrc 时代是全部)
