+++ app/shared/app-data-aidl/build.gradle.kts
@@ -8,9 +8,7 @@
  */
 
 plugins {
-    id("ani.base")
-    alias(libs.plugins.android.library)
-    id("kotlin-parcelize")
+    id("ani.android-library")
 }
 
 android {
