+++ build-logic/src/main/kotlin/ani.flatten-source-sets.gradle.kts
