+++ build-logic/src/main/kotlin/properties.kt
@@ -16,14 +16,8 @@ import org.gradle.kotlin.dsl.of
 import java.util.Properties
 
 /**
- * 读取 `local.properties` 的 [ValueSource].
- *
- * 为什么用 ValueSource 而不是直接 `File.readText()`:
- * Gradle 只把 ValueSource 的**返回值**登记为配置缓存输入, 文件内容真正变化时才会让缓存失效;
- * 而在配置期直接读文件 / 调用 `File.exists()` 会被登记成一堆零散的 file-system 输入,
- * 使缓存失效条件变得不可预测.
- *
- * [obtain] 必须足够快 —— 每次构建都会执行一次, 用来判断配置缓存能否复用.
+ * 读取 `local.properties`. 用 ValueSource 让 Gradle 只把返回值登记为配置缓存输入,
+ * 而不是配置期直接读文件产生的一堆零散 file-system 输入.
  */
 abstract class LocalPropertiesValueSource :
     ValueSource<Map<String, String>, LocalPropertiesValueSource.Parameters> {
@@ -33,8 +27,7 @@ abstract class LocalPropertiesValueSource :
 
     override fun obtain(): Map<String, String> {
         val file = parameters.localPropertiesFile.get().asFile
-        // 文件不存在就返回空. 历史实现会在这里 createNewFile(), 那是配置期副作用
-        // (它真的在仓库里留下过一个空的 buildSrc/local.properties).
+        // 不存在就返回空. 历史实现在这里 createNewFile(), 那是配置期副作用.
         if (!file.isFile) return emptyMap()
         val properties = Properties()
         file.inputStream().buffered().use(properties::load)
@@ -42,21 +35,13 @@ abstract class LocalPropertiesValueSource :
     }
 }
 
-/**
- * `local.properties` 的全部内容. 由 Gradle 缓存, 每次构建只解析一次.
- *
- * 用 `layout.settingsDirectory` 而不是 `rootProject.file(...)`, 避免跨项目访问.
- */
+/** `local.properties` 全部内容, 每次构建只解析一次. 用 settingsDirectory 避免跨项目访问. */
 val Project.localProperties: Provider<Map<String, String>>
     get() = providers.of(LocalPropertiesValueSource::class) {
         parameters.localPropertiesFile.set(layout.settingsDirectory.file("local.properties"))
     }
 
-/**
- * 按 `local.properties` -> `-D` 系统属性 -> 环境变量 -> Gradle property 的顺序查找.
- *
- * 全链路都是 provider, 因此每一个来源都会被登记为配置缓存输入.
- */
+/** 按 local.properties -> 系统属性 -> 环境变量 -> Gradle property 查找, 全链路 provider. */
 fun Project.aniProperty(name: String): Provider<String> =
     localProperties.map { it[name] }
         .orElse(providers.systemProperty(name))
@@ -68,7 +53,7 @@ fun Project.getProperty(name: String) =
 
 fun Project.getPropertyOrNull(name: String): String? =
     aniProperty(name).orNull
-    // extra 是项目本地的可变状态, 不是外部输入, 因此不需要 (也不能) 走 provider.
+    // extra 是项目本地可变状态, 不是外部输入, 不走 provider.
         ?: extensions.extraProperties.runCatching { get(name).toString() }.getOrNull()
 
 fun Project.getLocalProperty(key: String): String? = localProperties.get()[key]
