--- buildSrc/settings.gradle.kts
@@ -1,9 +1,0 @@
-rootProject.name = "ani"
-
-dependencyResolutionManagement {
-    versionCatalogs {
-        create("libs") {
-            from(files("../gradle/libs.versions.toml"))
-        }
-    }
-}
