+++ ci-helper/build.gradle.kts
@@ -14,8 +14,7 @@ import org.gradle.api.tasks.bundling.Zip
 import java.io.File
 
 plugins {
-    id("ani.base")
-    alias(libs.plugins.kotlin.jvm)
+    id("ani.jvm-library")
     alias(libs.plugins.kotlinx.atomicfu)
 }
