+++ datasource/bt/dmhy/dataset-tools/build.gradle.kts
@@ -10,8 +10,7 @@
 @file:Suppress("UnstableApiUsage")
 
 plugins {
-    id("ani.base")
-    alias(libs.plugins.kotlin.jvm)
+    id("ani.jvm-library")
     alias(libs.plugins.kotlin.plugin.serialization)
 }
