+++ tools/datasource-test-mcp/build.gradle.kts
@@ -8,8 +8,7 @@
  */
 
 plugins {
-    id("ani.base")
-    alias(libs.plugins.kotlin.jvm)
+    id("ani.jvm-library")
     alias(libs.plugins.kotlin.plugin.serialization)
     alias(libs.plugins.kotlin.plugin.compose)
     application
