+++ utils/build-config/build.gradle.kts
@@ -8,8 +8,7 @@
  */
 
 plugins {
-    id("ani.base")
-    alias(libs.plugins.android.library)
+    id("ani.android-library")
 }
 
 // TODO AGP Migration: Test new build config behavior
